Mark proposal operations as virtual. #644

+++ plugins/operation_history/include/golos/plugins/operation_history/history_object.hpp
@@ -52,7 +52,7 @@ namespace golos { namespace plugins { namespace operation_history {
         uint32_t block = 0;
         uint32_t trx_in_block = 0;
         uint16_t op_in_trx = 0;
-        uint64_t virtual_op = 0;
+        uint32_t virtual_op = 0;
         time_point_sec timestamp;
         buffer_type serialized_op;
     };
@@ -74,7 +74,7 @@ namespace golos { namespace plugins { namespace operation_history {
                     member<operation_object, uint32_t, &operation_object::block>,
                     member<operation_object, uint32_t, &operation_object::trx_in_block>,
                     member<operation_object, uint16_t, &operation_object::op_in_trx>,
-                    member<operation_object, uint64_t, &operation_object::virtual_op>,
+                    member<operation_object, uint32_t, &operation_object::virtual_op>,
                     member<operation_object, operation_id_type, &operation_object::id>>>,
             ordered_unique<
                 tag<by_transaction_id>,
